+++ dvc/stage/cache.py
@@ -217,7 +217,7 @@ class StageCache:
 
         from_fs = from_odb.fs
         to_fs = from_odb.fs
-        func = _log_exceptions(fs.utils.copy)
+        func = _log_exceptions(fs.generic.copy)
         runs = from_fs.path.join(from_odb.fs_path, "runs")
 
         ret = []
